% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-if-missing.R
\name{DefaultIfMising}
\alias{DefaultIfMising}
\alias{\%||\%}
\title{Default if Missing}
\usage{
x \%||\% y
}
\arguments{
\item{x}{input}

\item{y}{replacement value for 0 or NULL}
}
\value{
replacement value if x is NULL
}
\description{
Infix function will replace missing values with another value.
}
\examples{
## replace a NULL value with a replacement value
x <- NULL
x \%||\% "replacementValue"
print(x)

## replace a NULL value within a function with a replacement value
testFunction <- function(x=NULL, y = "bar"){
 x <-x \%||\% "foo"
print(c(x, y))
}
testFunction()


}
