% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-in.R
\name{\%nin\%}
\alias{\%nin\%}
\title{human readable title here. check similar packages.}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or NULL; the values to be matched}

\item{table}{vector of NULL; values to be matched against}
}
\value{
a logical vector
}
\description{
human readable title here. check similar packages.
}
